Let the game back out to the menu
Three faults on the level-exit path, found one behind the other.

The world was never released. gbh_EndLevel clears loadedMapObject_, which
is only the marker for *which* map was loaded; the vertex and index
buffers still held the last level and DrawWorld's only guard is that a
buffer exists. MapObject() goes null at the menu, so EnsureWorldLoaded
bailed out early and never replaced it, and the city stayed on screen -
and, under Remix, stayed in the scene being path traced. Renderer gains
ReleaseWorld, which drops the geometry and deliberately not the device,
window or tile textures; those outlive a level. It runs off a new
worldUploaded_ flag, because loadedMapObject_ has already been cleared by
the time the map goes. The ground sampler, the animated-tile tracking and
the camera's height easing are reset with it.

gbh_SetWindow takes floats. The game's only call site builds them with
fild/fstp at 0x004CAF1F and passes (0.0f, 0.0f, width-1, height-1);
declared as ints, 639.0f arrived as 1142312960 and NoteWindow threw it
out as out of range. So that call had never told the overlay anything,
and the front end was laid out to a default that happened to be right
until a level had been played and the camera struct had overwritten it.
The camera struct now only drives the overlay size while a world is
loaded - its pointer outlives the level it described, so at the menu it
is a freed allocation - and the overlay keeps gbh_SetWindow's last
rectangle to fall back to.

gbh_ConvertColour returns 5:6:5, which is what a renderer asked for
16-bit colour should return, and FlatRect was reinterpreting that as
ARGB8888 with an alpha forced on. Red landed in the byte 5:6:5 never
fills, green in the red byte, blue in the green. It is expanded properly
now.

Two theories are recorded as ruled out rather than left to be tried
again: gbh_BlitBuffer has zero call sites in gta2.exe, so nothing was
relying on a buffer restore being dropped, and gbh_CloseScreen is not in
the game's bound export table at all, so Shutdown was never firing on
level exit.

+++ dll/src/dllmain.cpp
@@ -597,9 +597,15 @@ __declspec(dllexport) void __stdcall gbh_SetAmbient(float ambient) {
     gta2dx9::LightsSetAmbient(ambient);
 }
 
-__declspec(dllexport) void __stdcall gbh_SetWindow(int a, int b, int c, int d) {
+// Floats, not ints. The game's only call site builds them with fild/fstp at
+// gta2.exe!0x004CAF1F and passes gbh_SetWindow(0.0f, 0.0f, width-1, height-1).
+// Declared as ints, 639.0f arrives as 1142312960, which NoteWindow rejected as
+// out of range - so this call had never once told the overlay anything, and the
+// menu was laid out to a default that happened to be right until a level had
+// been played.
+__declspec(dllexport) void __stdcall gbh_SetWindow(float a, float b, float c, float d) {
     if (Proxying() && g_p_gbh_SetWindow) {
-        Backend<void(__stdcall*)(int, int, int, int)>(g_p_gbh_SetWindow)(a, b, c, d);
+        Backend<void(__stdcall*)(float, float, float, float)>(g_p_gbh_SetWindow)(a, b, c, d);
         return;
     }
     g_world.GetOverlay().NoteWindow(a, b, c, d);
